backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
